Move the internal api to dedicated module
