--- docs/api/develop.md
@@ -1,37 +1,0 @@
-# Developer's API
-
-These APIs are for the internal usage and development of Spglib itself.
-
-## Python interface
-
-```{eval-rst}
-.. autofunction:: spglib.spglib.get_pointgroup
-```
-
-```{eval-rst}
-.. autofunction:: spglib.spglib.get_layergroup
-```
-
-```{eval-rst}
-.. autofunction:: spglib.spglib.get_symmetry_layerdataset
-```
-
-```{eval-rst}
-.. autofunction:: spglib.spglib.get_grid_point_from_address
-```
-
-```{eval-rst}
-.. autofunction:: spglib.spglib.get_stabilized_reciprocal_mesh
-```
-
-```{eval-rst}
-.. autofunction:: spglib.spglib.get_grid_points_by_rotations
-```
-
-```{eval-rst}
-.. autofunction:: spglib.spglib.get_BZ_grid_points_by_rotations
-```
-
-```{eval-rst}
-.. autofunction:: spglib.spglib.relocate_BZ_grid_address
-```
