+++ python/spglib/__init__.py
@@ -3,6 +3,9 @@
 # This file is part of spglib.
 # SPDX-License-Identifier: BSD-3-Clause
 
+# TODO: _internal should not be exposed, we only continue to do so for
+#  backwards compatibility
+from ._internal import *  # noqa: F403
 from ._version import __version__, __version_tuple__  # noqa: F401
 
 # fmt: off
@@ -13,30 +16,23 @@ from .spglib import (  # noqa: F401
     SpglibMagneticDataset,
     delaunay_reduce,
     find_primitive,
-    get_BZ_grid_points_by_rotations,
     get_error_message,
-    get_grid_point_from_address,
-    get_grid_points_by_rotations,
     get_hall_number_from_symmetry,
     get_ir_reciprocal_mesh,
-    get_layergroup,
     get_magnetic_spacegroup_type,
     get_magnetic_spacegroup_type_from_symmetry,
     get_magnetic_symmetry,
     get_magnetic_symmetry_dataset,
     get_magnetic_symmetry_from_database,
-    get_pointgroup,
     get_spacegroup,
     get_spacegroup_type,
     get_spacegroup_type_from_symmetry,
-    get_stabilized_reciprocal_mesh,
     get_symmetry,
     get_symmetry_dataset,
     get_symmetry_from_database,
     get_version,
     niggli_reduce,
     refine_cell,
-    relocate_BZ_grid_address,
     spg_get_commit,
     spg_get_version,
     spg_get_version_full,
